feat: Add apollo.config.js for Apollo client configuration
This commit adds the apollo.config.js file to the repository. The file contains the configuration for the Apollo client, specifying the service name and URL for the GraphQL API. This configuration enables the client to connect to the "sg-bus" service at "http://localhost:3000/api/graphql". The apollo.config.js file is included in the TypeScript compilation by updating the tsconfig.json file. This change ensures that the Apollo client is properly configured and can communicate with the GraphQL API.