feat: Add getTimeUntilArrival function to calculate time difference
This commit adds the getTimeUntilArrival function to the timeUtil.ts file. The function calculates the time difference between the estimated arrival time and the current time. If the time difference is less than or equal to 60 seconds, it returns "Arr". Otherwise, it returns the number of minutes until arrival. This function will be used to display the estimated time until arrival for buses.