refactor: Update BusArrivalInfo to use dynamic refresh interval
The BusArrivalInfo component has been updated to use a dynamic refresh interval for fetching bus arrival data. Instead of a fixed interval of 3000 milliseconds, the refresh interval is now determined by the value of the NEXT_PUBLIC_BUS_ARRIVAL_REFRESH_INTERVAL environment variable, with a default value of 15000 milliseconds. This change allows for more flexibility in controlling the frequency of data updates.